requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==0.7
Django==4.2.6
django-ckeditor==6.7.0
django-crispy-forms==2.0
django-filter==23.3
django-js-asset==2.1.0
django-seed==0.3.1
djangorestframework==3.14.0
export==0.2.1
factory-boy==3.2.1
Faker==19.8.0
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
Markdown==3.4.4
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.0
openai==0.28.1
packaging==23.2
pandas==2.1.1
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.11.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
toposort==1.10
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
yarl==1.9.2
zipp==3.17.0
